+++ packages/authhero/test/authentication-flows/common.spec.ts
@@ -1,6 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { Context } from "hono";
 import {
+  createAuthResponse,
   createAuthTokens,
   createSession,
 } from "../../src/authentication-flows/common";
@@ -92,6 +93,76 @@ describe("common", () => {
         token_type: "Bearer",
         expires_in: 86400,
       });
+    });
+  });
+
+  it("should create a code when the response type is code", async () => {
+    const { env } = await getTestServer();
+    const ctx = {
+      env,
+      var: {
+        tenant_id: "tenantId",
+      },
+    } as Context<{
+      Bindings: Bindings;
+      Variables: Variables;
+    }>;
+
+    // Create the login session and code
+    const loginSession = await env.data.logins.create("tenantId", {
+      expires_at: new Date(Date.now() + 1000 * 60 * 5).toISOString(),
+      authParams: {
+        client_id: "clientId",
+        username: "foo@example.com",
+        scope: "",
+        audience: "http://example.com",
+        redirect_uri: "http://example.com/callback",
+      },
+    });
+
+    const client = await env.data.clients.get("clientId");
+    const user = await getPrimaryUserByEmail({
+      userAdapter: env.data.users,
+      tenant_id: "tenantId",
+      email: "foo@example.com",
+    });
+
+    if (!client || !user) {
+      throw new Error("Client or user not found");
+    }
+
+    const autResponse = await createAuthResponse(ctx, {
+      authParams: {
+        client_id: "clientId",
+        response_type: AuthorizationResponseType.CODE,
+        scope: "openid",
+        redirect_uri: "http://example.com/callback",
+      },
+      client,
+      user,
+      loginSession,
+    });
+
+    expect(autResponse.status).toEqual(302);
+    const location = autResponse.headers.get("location");
+    if (!location) {
+      throw new Error("No location header");
+    }
+
+    console.log(location);
+
+    const redirectUri = new URL(location);
+    const codeQuerystring = redirectUri.searchParams.get("code");
+
+    const code = await env.data.codes.get(
+      "tenantId",
+      codeQuerystring!,
+      "authorization_code",
+    );
+    expect(code).toMatchObject({
+      code_id: codeQuerystring,
+      code_type: "authorization_code",
+      user_id: "email|userId",
     });
   });